OCR/requirements.txt
opencv-python
pytesseract
numpy
nltk
pillow
imutils
dash
dash-extensions
matplotlib
flask
flask-cors
